Fractional gdiff: continuous diffusion knob, with values < 1 supported
gdiff is now a float instead of int. floor(gdiff) full 3×3 box-blur
passes are still applied, then one partial pass with strength s =
gdiff - floor(gdiff). The partial pass is:

  F'(x) = (1 - s) * F(x) + s * mean_3x3(F(x))
        = F(x) - s * (F(x) - mean)

s = 1 collapses to the existing diffuse_food_once, s = 0 is identity.
So gdiff = 0.5 puts F(x) halfway between its current value and its
local 3×3 mean — half the diffusion of a single full pass — and lets
the user dial diffusion below the previous integer minimum of 1.

Touches:
- evoca.c: ggdiff is float; new diffuse_food_partial(s) helper; step
  loop does floor(ggdiff) full passes then one partial if frac > 0.
- evoca.h: signature change for evoca_set_gdiff / evoca_get_gdiff.
- evoca_py.py: ctypes signatures use c_float; gdiff stored as float.
- controls.py: gdiff slider is FloatSlider with step 0.01.

Verified mass-conserving and matching analytical predictions for
gdiff ∈ {0, 0.25, 0.5, 1.0, 2.5} on a delta-function input.

+++ C/evoca.h
@@ -43,7 +43,7 @@ void evoca_free(void);
 
 void evoca_set_food_inc(float f);
 void evoca_set_m_scale(float m);
-void evoca_set_gdiff(int d);
+void evoca_set_gdiff(float d);
 void evoca_set_mu_lut(float m);
 void evoca_set_mu_egenome(float m);
 void evoca_set_tax(float t);
@@ -183,7 +183,7 @@ uint8_t *evoca_get_births(void); /* [N*N] 0=none, 1=birth, 2=mutant birth */
 uint8_t *evoca_get_alive(void);  /* [N*N] 1=alive organism, 0=dead slot */
 int      evoca_get_N(void);
 int      evoca_get_cell_px(void);
-int      evoca_get_gdiff(void);
+float    evoca_get_gdiff(void);
 float    evoca_get_mu_lut(void);
 float    evoca_get_mu_egenome(void);
 float    evoca_get_tax(void);
